+++ src/unix/blit/blit_yuy2.h
@@ -31,7 +31,7 @@ INLINE void FUNC_NAME(blit_normal_line_2)(SRC_PIXEL *src, SRC_PIXEL *end,
 }
 
 BLIT_BEGIN(blit_normal)
-  switch(sysdep_display_params.widthscale)
+  switch((int)sysdep_display_params.widthscale)
   {
     case 1:
       BLIT_LOOP(blit_normal_line_1, 1);
@@ -43,7 +43,7 @@ BLIT_BEGIN(blit_normal)
 BLIT_END
 
 BLIT_BEGIN(blit_fakescan_h)
-  switch(sysdep_display_params.widthscale)
+  switch((int)sysdep_display_params.widthscale)
   {
     case 1:
       BLIT_LOOP(blit_normal_line_1, 2);
